separate recursive partitioning, rename to deep partitioning
