+++ mt-kahypar/partition/deep_partitioning.h
@@ -2,6 +2,7 @@
  * This file is part of KaHyPar.
  *
  * Copyright (C) 2019 Tobias Heuer <tobias.heuer@kit.edu>
+ * Copyright (C) 2021 Nikolai Maas <nikolai.maas@student.kit.edu>
  *
  * KaHyPar is free software: you can redistribute it and/or modify
  * it under the terms of the GNU General Public License as published by
@@ -21,30 +22,14 @@
 #pragma once
 
 #include "mt-kahypar/partition/context.h"
-#include "mt-kahypar/partition/initial_partitioning/i_initial_partitioner.h"
+
 
 namespace mt_kahypar {
+namespace deep_partitioning {
 
-class RecursiveInitialPartitioner: public IInitialPartitioner {
- private:
+PartitionedHypergraph partition(Hypergraph& hypergraph, const Context& context);
 
-  static constexpr bool enable_heavy_assert = false;
+void partition(PartitionedHypergraph& hypergraph, const Context& context);
 
-
-
- public:
-  RecursiveInitialPartitioner(PartitionedHypergraph& hypergraph, const Context& context);
-  RecursiveInitialPartitioner(const RecursiveInitialPartitioner&) = delete;
-  RecursiveInitialPartitioner(RecursiveInitialPartitioner&&) = delete;
-  RecursiveInitialPartitioner & operator= (const RecursiveInitialPartitioner &) = delete;
-  RecursiveInitialPartitioner & operator= (RecursiveInitialPartitioner &&) = delete;
-
- private:
-  void initialPartitionImpl() final ;
-
- private:
-  PartitionedHypergraph& _hg;
-  const Context& _context;
-};
-
+}  // namespace deep_partitioning
 }  // namespace mt_kahypar
